chore: migrate to ESLint 9 flat config
