--- frontend/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "extends": ["next/core-web-vitals", "plugin:@typescript-eslint/recommended", "prettier"]
-}
