+++ frontend/.gitignore
@@ -11,6 +11,7 @@
 # next.js
 /.next
 /out
+**/next-env.d.ts
 
 # production
 /build
@@ -35,5 +36,3 @@ yarn-error.log*
 
 # vercel
 .vercel
-.npmrc
-.yarnrc
