+++ frontend/eslint.config.mjs
@@ -1,0 +1,22 @@
+import { FlatCompat } from '@eslint/eslintrc';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import js from '@eslint/js';
+import tseslint from 'typescript-eslint';
+import prettier from 'eslint-config-prettier/flat';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+const compat = new FlatCompat({ baseDirectory: __dirname });
+
+const config = [
+  {
+    ignores: ['.next/**', 'node_modules/**', 'cypress/**', 'coverage/**', '.jest/**', '**/*.test.*', 'next-env.d.ts'],
+  },
+  js.configs.recommended,
+  ...tseslint.configs.recommended,
+  ...compat.extends('next/core-web-vitals'),
+  prettier,
+];
+
+export default config;
