+++ frontend/next.config.mjs
@@ -1,0 +1,43 @@
+import envalid from 'envalid';
+import bundleAnalyzer from '@next/bundle-analyzer';
+
+const authDependent = envalid.makeValidator((x) => {
+  const authEnabled = process.env.HEALTH_AUTH === 'true';
+
+  if (authEnabled && !x.length) {
+    throw new Error(`Can't be empty if "HEALTH_AUTH" is true`);
+  }
+
+  return x;
+});
+
+envalid.cleanEnv(process.env, {
+  NEXT_PUBLIC_API_URL: envalid.str(),
+  HEALTH_AUTH: envalid.bool(),
+  HEALTH_USERNAME: authDependent(),
+  HEALTH_PASSWORD: authDependent(),
+});
+
+const withBundleAnalyzer = bundleAnalyzer({ enabled: process.env.ANALYZE === 'true' });
+
+export default withBundleAnalyzer({
+  //output: 'standalone',
+  i18n: {
+    locales: ['sv'],
+    defaultLocale: 'sv',
+  },
+  images: {
+    domains: [process.env.DOMAIN_NAME],
+    formats: ['image/avif', 'image/webp'],
+  },
+  basePath: process.env.BASE_PATH,
+  sassOptions: {
+    prependData: `$basePath: '${process.env.BASE_PATH}';`,
+  },
+  experimental: {
+    optimizePackageImports: ['lucide-react', '@sk-web-gui/react'],
+  },
+  async rewrites() {
+    return [{ source: '/napi/:path*', destination: '/api/:path*' }];
+  },
+});
