--- frontend/tailwind.config.js
@@ -1,26 +1,0 @@
-module.exports = {
-  mode: 'jit',
-  content: [
-    './src/pages/**/*.{js,ts,jsx,tsx}',
-    './src/components/**/*.{js,ts,jsx,tsx}',
-    './src/layouts/**/*.{js,ts,jsx,tsx}',
-    './node_modules/@sk-web-gui/*/dist/**/*.{js,mjs}',
-  ],
-  darkMode: 'class', // or 'media' or 'class'
-  theme: {
-    /* screens: {
-      sm: '640px',
-      md: '768px',
-      lg: '1024px',
-    },*/
-  },
-  plugins: [
-    //require('@tailwindcss/typography'),
-    require('@tailwindcss/forms'),
-    require('@tailwindcss/container-queries'),
-    require('@sk-web-gui/core')({
-      colors: [],
-      cssBase: true,
-    }),
-  ],
-};
